python/requirements.txt
PyQt6
opencv-python
torch
numpy
Pillow
minerl
pytest